Mod: `js_denotest/util` renamed to `/test-util` - for `static/util` tests organize
