+++ src/tests/js_denotest/packet_conv/audio.test.js
@@ -13,8 +13,8 @@ import { assertEquals, assertThrows } from "jsr:@std/assert@1";
 
 import { ONE_FRAME_SAMPLES } from "../../../static/AUDIO_PARAM.js";
 import { packet_audio_encode, packet_audio_decode, is_audio_packet, AUDIO_PACKET_TYPE_ID, SILENT_AUDIO_PACKET_TYPE_ID } from "../../../static/packet_conv/audio.js";
-import { generate_rand_float32array } from "../util/rand_f32a.js";
-import { assertAlmostEqualsArray } from "../util/assertAlmostEqualsArray.js";
+import { generate_rand_float32array } from "../test-util/rand_f32a.js";
+import { assertAlmostEqualsArray } from "../test-util/assertAlmostEqualsArray.js";
 
 
 const ERR_INT16_AND_FLOAT32 = 1 / 32767;
