+++ src/tests/js_denotest/test-util/assertThrows_ecma.js
